Renamed SignupForm.svelte to login.svelte
